wipe gradients on unpolish, gcc mourns
git-svn-id: https://cloudcity.svn.sourceforge.net/svnroot/cloudcity@367 d23db356-9f35-0410-88be-f4158e15d4e5

+++ input.cpp
@@ -121,7 +121,7 @@ drawSBArrow(QStyle::SubControl sc, QPainter *painter, QStyleOptionSpinBox *optio
         }
 
         bool isEnabled = option->stepEnabled & sef;
-        bool hover = isEnabled && (option->activeSubControls == sc);
+        bool hover = isEnabled && (option->activeSubControls == (int)sc);
         bool sunken = hover && (option->state & QStyle::State_Sunken);
